+++ src/main/java/nextstep/web/repository/database/ReservationJdbcRepository.java
@@ -1,10 +1,11 @@
-package nextstep.web.repository;
+package nextstep.web.repository.database;
 
 import lombok.RequiredArgsConstructor;
 import nextstep.domain.Reservation;
 import nextstep.domain.Theme;
 import nextstep.web.exception.BusinessException;
 import nextstep.web.exception.CommonErrorCode;
+import nextstep.web.repository.ReservationRepository;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
