+++ src/test/java/nextstep/web/repository/ReservationJdbcRepositoryTest.java
@@ -3,6 +3,7 @@ package nextstep.web.repository;
 import nextstep.domain.Reservation;
 import nextstep.domain.Theme;
 import nextstep.web.exception.BusinessException;
+import nextstep.web.repository.database.ReservationJdbcRepository;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
